requirements.txt
pytest
allure-pytest
selenium
webdriver-manager